Add assume role support
Add support to assume IAM roles when performing CLI operations. It's
possible to specify a role ARN that should be attempted to be assumed.
If the specified role required MFA, this can also be provided using
the --iam-mfa and --iam-code flags. If the --iam-code flag is not
specified, then a code will be prompted for.

A role ARN is typically on the format:
	arn:aws:iam::<account ID>:role/<role name>
An MFA device ARN is typically on the format:
	arn:aws:iam::<account ID>:mfa/<username>

The code prompt is printed to stderr, so it should not interfere with
the output result in stdout.