Add context to handle cleanup
A major refactoring to properly use context. This enables cleanup
of resources in a very elegant way and works well with the AWS SDK
to make sure that no more work than nessecary is ever performed.

If something fails or the user interupts the execution (e.g. with
^C), any ongoing request to AWS will be cancelled, and temporary
resources will be cleaned up.

Cleanup should be super fast, but is in-fact blocking. So in order
to exit the CLI during this phase a SIGTERM or SIGKILL is required.