Fix channel race condition bug
Fixes a bug where if initializing a new service fails, there is
a potential risk for a panic to happen as the goroutine still
continues execution after writing the error to the result channel.

This meant that even though an error is written to the result
channel, the subsequent function call still tries to use the nil
service before the result channel has been handled in the select
statement.