Improve tagging on temporary resources
Make sure that all temporary resources created by the CLI are
properly tagged so that they can be identified and controlled
through IAM policy conditions.

Instances, volumes and snapshots created by the CLI will have
the tag "metavisor-cli: true" on them.

This also adds a template IAM policy with all the minimum
requirements for the CLI to work properly. See the
policy_template.json file for this.