Add easier install instructions to README
Add more details about how to install the CLI, to make it easier for
new users to get started. This includes separating install
instructions between installing from source code and from binary.

It also adds some examples on how to easily install the binaries
on Linux, macOS and Windows. Although the steps for Windows are
a bit limited.