Make region flag optional for wrap-instance
If no region is specified for the wrap instance command, the CLI
will try to figure it out. This is possible since the instance ID
should be locally unique across regions within the current account,
especially for a limited time frame (only looking at running
instances at the time of the command being executed).