Better handling of missing AWS creds
Make sure that if there are no valid AWS credentials present,
there will be an error message that makes sense (rather than
just showing the internal AWS error).

The error happens if there are no AWS credential environment
variables set or no config in ~/.aws/config (or if there are
credentials but they are invalid).