Improve error handling
Fixes a typo in an error message. Also gives a better error message
when creating a wrapped AMI fails. It tries to look at the failure
and give a more actionable error message. In most cases where
the instance shuts down right after getting ready it's because
the MV can't authorize.